+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.2.0] 2021-01-03
 ### Changes
+- New node for registration via voice call.
 - Save all account info in sub directory "signal" to avoid conflicts. Reregistration is neccessary.
 
 ## [1.1.0] 2020-12-31